HARMONY-842: Make sure all tasks are using node16 in their Dockerfiles. Refactor work-item interface into a separate file so that service runner does not need database libraries as dependencies.
