+++ app/workers/work-failer.ts
@@ -1,11 +1,12 @@
 import { Logger } from 'winston';
-import { WorkItemStatus, getWorkItemsByAgeAndStatus, updateWorkItemStatuses } from '../models/work-item';
+import { getWorkItemsByAgeAndStatus, updateWorkItemStatuses } from '../models/work-item';
 import env from '../util/env';
 import { Worker } from './worker';
 import db, { Transaction } from '../util/db';
 import sleep from '../util/sleep';
 import { Job, JobStatus } from '../models/job';
 import { completeJob } from '../util/job';
+import { WorkItemStatus } from '../models/work-item-interface';
 
 export interface WorkFailerConfig {
   logger: Logger;
