+++ package.json
@@ -102,7 +102,7 @@
     "shpjs": "^4.0.2",
     "simple-oauth2": "git+https://github.com/bilts/simple-oauth2.git#fix-encoding",
     "source-map-support": "^0.5.19",
-    "sqlite3": "^4.2.0",
+    "sqlite3": "^5.0.3",
     "swagger-ui-express": "^4.1.4",
     "tmp": "^0.2.1",
     "tmp-promise": "^3.0.3",
