+++ tasks/query-cmr/Dockerfile
@@ -1,4 +1,4 @@
-FROM node:12-alpine
+FROM node:16-alpine
 
 WORKDIR /app
 COPY built package.json package-lock.json ./
