+++ tasks/service-runner/Dockerfile
@@ -1,8 +1,7 @@
-FROM node:12-alpine
+FROM node:16-alpine
 
 RUN apk add bash vim curl git
 RUN git config --global url."https://".insteadOf ssh://
-RUN npm install -g npm@7
 
 RUN mkdir -p /service-runner
 RUN mkdir -p /tmp/metadata
