+++ tasks/service-runner/app/service/service-runner.ts
@@ -4,7 +4,7 @@ import stream from 'stream';
 import { sanitizeImage } from '../../../../app/util/string';
 import env from '../util/env';
 import logger from '../../../../app/util/log';
-import WorkItem from '../../../../app/models/work-item';
+import { WorkItemRecord } from '../../../../app/models/work-item-interface';
 import axios from 'axios';
 
 // Must match where harmony expects artifacts in workflow-orchestration.ts
@@ -89,7 +89,7 @@ function _getErrorMessage(logStr: string, catalogDir: string): string {
   * @param operation - The requested operation
   * @param callback - Function to call with result
   */
-export async function runQueryCmrFromPull(workItem: WorkItem): Promise<ServiceResponse> {
+export async function runQueryCmrFromPull(workItem: WorkItemRecord): Promise<ServiceResponse> {
 
   const { operation, scrollID } = workItem;
   const catalogDir = `${ARTIFACT_DIRECTORY}/${operation.requestId}/${workItem.id}/outputs`;
@@ -129,7 +129,7 @@ export async function runQueryCmrFromPull(workItem: WorkItem): Promise<ServiceRe
  * @param operation - The requested operation
  * @param callback - Function to call with result
  */
-export async function runServiceFromPull(workItem: WorkItem): Promise<ServiceResponse> {
+export async function runServiceFromPull(workItem: WorkItemRecord): Promise<ServiceResponse> {
   try {
     const { operation, stacCatalogLocation } = workItem;
     // support invocation args specified with newline separator or space separator
