+++ tasks/service-runner/package.json
@@ -29,13 +29,10 @@
     "date-fns": "^2.25.0",
     "dotenv": "^10.0.0",
     "express": "^4.17.1",
-    "knex": "^0.95.14",
-    "knex-paginate": "^2.3.0",
     "lodash": "^4.17.21",
     "lodash.camelcase": "^4.3.0",
     "node-fetch": "^2.6.1",
     "python-shell": "^3.0.1",
-    "sqlite3": "^4.2.0",
     "superagent": "^6.1.0",
     "tmp": "^0.2.1",
     "winston": "^3.7.2"
