+++ test/util/job.ts
@@ -5,10 +5,11 @@ import { buildWorkItem } from '../helpers/work-items';
 import { hookTransaction } from '../helpers/db';
 import { cancelAndSaveJob } from '../../app/util/job';
 import { JobStatus } from '../../app/models/job';
-import { getWorkItemsByJobId, WorkItemStatus } from '../../app/models/work-item';
+import { getWorkItemsByJobId } from '../../app/models/work-item';
 import db from '../../app/util/db';
 import log from '../../app/util/log';
 import { hookClearScrollSessionExpect } from '../helpers/hooks';
+import { WorkItemStatus } from '../../app/models/work-item-interface';
 
 const aTurboJob = buildJob({ username: 'doe' });
 const firstTurboWorkItem = buildWorkItem({ jobID: aTurboJob.jobID });
